Emby Logos and Themes changed to Jellyfin
The splash screen, in app logos, and launch icon logos have been changed to Jellyfin's logo which was taken directly from the img folder from Jellyfin web. The user accounts icon and rotten tomatos ratings have been updated as well as the player controls and slider being changed to the Jellyfin purple insdead of the emby green. The name of the app was changed to Jellyfin in various locations except for the credit page as I thought it better to leave the names of the original contributers and emby's website.

No major code was changed and I understand that this is an old app for older samsung TVs (like my own), but I as I am still learning C# I wanted to start contributing in a much smaller area like this one.